Fix up foreign key from videos to announcements
All videos should be attached to an announcement from the same bucket,
so make the schema enforce that. One could probably argue that this means
that bucket_slug doesn't need to exist in the videos table but in my
opinion one would be dead wrong :) - the bucket is an integral part of the
video so while it's sort of duplication, it isn't denormalisation.