Skeleton HTML files for static view interface
My intention is to use the id="main-content" as a merge point - layout.html
will be applied to the per-page HTML files.

The announcement thing is a bit clumsily worded because I really want to
avoid having to handle quant/pluralisation for the first pass - that would
raise all sorts of fascinating questions wrt using i18n stuff for _quant and
similar, and, well, blow that for a game of soliders.

Mark explained what was required for the FLV player, but (1) since I try and
avoid knowing anything about flash I had no idea what he was talking about,
(2) the first deployment will probably be on dreamhost who have their own
player and so while that's going to be all we support to begin with I'd like
to make it reasonably easy for somebody to write support for something else.

I consider this a good excuse to completely ignore it for the moment.