Iniitial IdiotBox.pm and script/idiotbox
Basic dispatch is in for the first bits - front page, bucket and video pages.

Not sure whether I'll try and make IdiotBox itself a Moose class or not yet.

Blocking out the structure for rendering and writing a quick bit of code to
generate a layout zoom; this is just a sketch though since Zoom will need some
updating before I can try and run it.